Removes type cast warnings.

+++ pkzo/Config.h
@@ -85,10 +85,10 @@ namespace pkzo
     inline
     bool iequals(const std::string& a, const std::string& b)
     {
-        unsigned int sz = a.size();
+        size_t sz = a.size();
         if (b.size() != sz)
             return false;
-        for (unsigned int i = 0; i < sz; ++i)
+        for (size_t i = 0; i < sz; ++i)
             if (tolower(a[i]) != tolower(b[i]))
                 return false;
         return true;
